+++ releasenotes.md
@@ -19,6 +19,7 @@
 - Add `stack-probe` and `stack-protector` project options. #3437
 - Add `@stackprobe(level)`, `@nostackprobe`, `@stackprotector(level)` and `@nostackprotector` attributes on functions and lambdas. #3437
 - Naked functions no longer have stack protectors by default.
+- Add `@noredzone` attribute.
 
 ### Stdlib changes
 - `CachedInStream` and `CachedOutStream` added.